Rules now have a new field, and function
The new field has type RuleType_t
and is used to define what type the rule is.
This can be used by lexers, parsers, ast generators
to produce more performant code with less duplication

There is also a IsEmpty function.
This function is used to check if the rule is empty or not