Update generation
Files aren't ignored now if they are empty.
It might occur that Deamer wants to generate a file, but the file was empty.
This resulted Deamer into thinking that it may not be used anymore.

Bison generation now adds Flex as a target to be compiled first.
This solves the problem where cmake compiled in incorrect order.