Add special LDO "UnInitialized"
The UnInitialized LDO is used for uninitialized LDO's.
Whenever you want to specify something that is not implemented,
you can use this LDO.

The threat analyzer will give a meaningful error message showing that a given
reference is not properly initialized.

DLDL will use this class to indicate unknown references.